requirements.txt
diffusers==0.27.0
transformers==4.40.2
peft==0.10.0
accelerate==0.30.1
torch>=2.3.0
torchvision
gpustat
albumentations
hydra-core
rich
tensorboard
torchmetrics
timesformer
